Move dynamic to models
