+++ app/Decsys/Data/Entities/Mongo/Webhook.cs
@@ -1,3 +1,5 @@
+using Decsys.Models.Webhooks;
+
 namespace Decsys.Data.Entities.Mongo;
 
 public class Webhook
@@ -11,23 +13,3 @@ public class Webhook
     public Dictionary<string, string> TriggerParameters { get; set; } = new Dictionary<string, string>();
     public List<TriggerFilter> TriggerFilters { get; set; } = new List<TriggerFilter>();
 }
-
-public class TriggerFilter
-{
-    public string Name { get; set; } = string.Empty;
-    public string Value { get; set; } = string.Empty;
-    public FilterOperator Operator { get; set; }
-}
-
-public enum FilterOperator
-{
-    EQUALS,
-    NOT_EQUALS,
-}
-
-public enum EventType
-{
-    PAGE_NAVIGATION,
-    SURVEY_COMPLETE
-}
-
