+++ app/Decsys/Models/Webhooks/WebhookModel.cs
@@ -1,5 +1,3 @@
-using Decsys.Data.Entities.Mongo;
-
 namespace Decsys.Models.Webhooks;
 
 public class WebhookModel
@@ -11,4 +9,23 @@ public class WebhookModel
     public EventType EventType { get; set; }
     public Dictionary<string, string> TriggerParameters { get; set; } = new Dictionary<string, string>();
     public List<TriggerFilter> TriggerFilters { get; set; } = new List<TriggerFilter>();
+}
+
+public class TriggerFilter
+{
+    public string Name { get; set; } = string.Empty;
+    public string Value { get; set; } = string.Empty;
+    public FilterOperator Operator { get; set; }
+}
+
+public enum FilterOperator
+{
+    EQUALS,
+    NOT_EQUALS,
+}
+
+public enum EventType
+{
+    PAGE_NAVIGATION,
+    SURVEY_COMPLETE
 }
